requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
sentence-transformers
scikit-learn
numpy
torch
mysql-connector-python